+++ src/net/nexttext/behaviour/control/Delay.java
@@ -40,7 +40,7 @@ public class Delay extends AbstractAction {
      * 
      * @param duration In seconds
      */
-    public Delay( Action action, double duration ) {
+    public Delay( Action action, float duration ) {
         this.action = action;
         this.timeLeft = 1000*(long)duration;
         properties().init("Duration", new NumberProperty(1000*duration));
